requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.9
